refactor: move composition to root directory
