+++ desktopApp/src/jvmMain/kotlin/main.kt
@@ -2,7 +2,7 @@ import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import com.akjaw.ai.assistant.shared.App
 import com.akjaw.ai.assistant.shared.MainView
-import com.akjaw.ai.assistant.shared.chat.composition.Dependencies
+import com.akjaw.ai.assistant.shared.composition.Dependencies
 import com.akjaw.ai.assistant.shared.chat.data.database.ProductionDriverFactory
 import com.akjaw.ai.assistant.shared.chat.data.database.createDatabase
