+++ shared/src/commonMain/kotlin/com/akjaw/ai/assistant/shared/dashboard/ui/Dashboard.kt
@@ -30,7 +30,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.akjaw.ai.assistant.shared.chat.composition.Dependencies
+import com.akjaw.ai.assistant.shared.composition.Dependencies
 import com.akjaw.ai.assistant.shared.chat.domain.ChatMessageHandler
 import com.akjaw.ai.assistant.shared.chat.domain.createKtorClient
 import com.akjaw.ai.assistant.shared.chat.presentation.ChatScreenStateHolder
